requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

# Machine Learning
torch==2.1.0
transformers==4.34.0
sentencepiece==0.1.99
tokenizers==0.14.0
timm==0.9.5
Pillow==10.0.0

# Database & Cache
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.1

# Utilities
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0

# Documentation
swagger-ui-bundle==0.9.3
python-multipart==0.0.6 